backend/requirements.txt
# Core Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Environment & Configuration
python-dotenv==1.0.0

# Database (SQLite)
SQLAlchemy==2.0.23

# Authentication
firebase-admin==6.3.0

# File Processing
Pillow==10.1.0

# Production Server
gunicorn==21.2.0